requirements.txt
telethon
piexif
pillow
pillow_heif
